requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
twilio==6.35.1
urllib3==1.25.7
